pyproject.toml
[tool.poetry]
name = "agentic-patterns"
version = "1.0.0"
description = ""
authors = ["MichaelisTrofficus <miguel.otero.pedrido.1993@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
colorama = "^0.4.6"
graphviz = "^0.20.3"
deepseek = "^1.0.0"
jupyter = "^1.0.0"
python-dotenv = "^1.0.1"
types-colorama = "^0.4.15.20240311"
requests = "^2.32.4"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
black = "^24.8.0"
reorder-python-imports = "^3.13.0"
pre-commit-hooks = "^4.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
